+++ benchmarking/loop/node_setup.sh
